pyproject.toml
[project]
name = "sql-query-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.20",
    "crewai>=0.186.1",
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.27",
    "llama-index>=0.14.0",
    "loguru>=0.7.3",
    "mysql-connector-python>=9.4.0",
    "ollama>=0.5.4",
    "openai>=1.107.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pyodbc>=5.2.0",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.0",
    "snowflake-connector-python>=3.17.3",
    "sqlalchemy>=2.0.43",
    "sqlparse>=0.5.3",
    "streamlit>=1.49.1",
    "tqdm>=4.67.1",
    "transformers>=4.56.1",
    "uvicorn>=0.35.0",
]
